Updated to not use objectid as branch name.
This is to help with hudson/jenkins integration where you cannot do a
checkout of master with the master branch as the name. So this way if
it is unable to resolve a branch name, it defaults to not doing any
version alterations at all instead of to the object id.