Added MOJO to install modified pom file
A limitation of the existing git version mojo is that it modified the
version number of the attached artifacts and deployed them correctly
but it did not go through and modify the actual poms to reflect those
said versions. Thus for anything that was a reactorized project which
had an artifact depending on another artifact within the reactor, the
version number it was told to fetch was not the modified version number
causing the artifact to not be retrievable downstream.

This often resulted in NullPointerException in downstream projects like
when assembling a package with dependencies on above reactorized lib.