Resolved situation where superpom is outside git
Certain situations people will make a superpom to reactorize multiple
projects together. This resolves both the issue of each subproject is a
different git project then the superpom holder and the issue of where
the superpom is not a git project to began with.

Essentially, if this plugin is called and the project is not a git repo
then it will no longer error out but instead will default to the
original version number.